add static networking parameters to razor policies
This is a stop-gap until we can upgrade to a newer version of
razor. Include static networking parameters in the policies and bind
them to their nodes. Use this data in the ESXi ks.cfg.erb to configure
a static management network address.

WARNING: this patch currently short-circuits password decryption for
initial testing.