ASM-8310 Call windows finished url after reboot
Previously the Windows installer called the "finished" stage_done_url
at the end of second-stage.ps1 However the installer will still reboot
twice after that before Windows is really up and ready to go.

This change adds a [SetupComplete.cmd][0] which Windows will automatically
execute after reboot. The finished url is called in the script.

The code to download and save the SetupComplete.cmd was cribbed from
the upstream razor-server project. They use that script for puppet
agent broker installation.

[0] https://technet.microsoft.com/en-us/library/cc766314(v=ws.10).aspx